refactor #1 fixed notifier logic
