+++ src/pages/content/chatgpt-notifier.ts
@@ -1,7 +1,9 @@
-// TODO: #1 REFACTOR
-import type { backgroundRoutes } from "../background/routes";
+import { backgroundRoutes } from "../background/routes";
 import { createRuntimeRpcClient } from "../lib/rpc/client";
 
+const callRuntimeRpc = createRuntimeRpcClient<typeof backgroundRoutes>();
+
+// TODO: #1 １個でいい
 const ANCHOR_SELECTORS = [
   "#thread-bottom-container",
   "#thread-bottom",
@@ -11,7 +13,6 @@ const ANCHOR_SELECTORS = [
 const STOP_BUTTON_SELECTOR = 'button[data-testid="stop-button"]';
 const NOTIFY_COOLDOWN_MS = 1500;
 const HEALTHCHECK_INTERVAL_MS = 2000;
-const callRuntimeRpc = createRuntimeRpcClient<typeof backgroundRoutes>();
 
 function resolveAnchor(): Element | null {
   for (const selector of ANCHOR_SELECTORS) {
@@ -27,7 +28,7 @@ function isGenerating(): boolean {
 
 function notifyFinished(): void {
   void callRuntimeRpc({
-    name: "background.notify",
+    name: "common.notify",
     title: "ChatGPT",
     message: `${document.title} で回答生成が終了しました`,
   });
